# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RedPath Sentinel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardStats {
  totalHosts: number;
  totalOpenPorts: number;
  highRiskServices: number;
  identityRisks: number;
  easyTargets: number;
  mediumTargets: number;
  hardTargets: number;
  totalAttackPaths: number;
}

export interface Scan {
  id: number;
  name: string;
  uploadedAt: string;
  hostCount: number;
  /** @nullable */
  rawXml?: string | null;
}

export interface ScanInput {
  name: string;
  xmlContent: string;
}

export interface ServiceInfo {
  port: number;
  protocol: string;
  serviceName: string;
  /** @nullable */
  product?: string | null;
  /** @nullable */
  version?: string | null;
  /** @nullable */
  riskLevel?: string | null;
}

export type WeaknessRiskLevel = typeof WeaknessRiskLevel[keyof typeof WeaknessRiskLevel];


export const WeaknessRiskLevel = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface Weakness {
  port: number;
  serviceName: string;
  riskLevel: WeaknessRiskLevel;
  observation: string;
  whyItMatters: string;
  recommendation: string;
}

export type HostDifficultyScore = typeof HostDifficultyScore[keyof typeof HostDifficultyScore];


export const HostDifficultyScore = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export type HostHostType = typeof HostHostType[keyof typeof HostHostType];


export const HostHostType = {
  'domain-controller': 'domain-controller',
  server: 'server',
  workstation: 'workstation',
  unknown: 'unknown',
} as const;

export interface Host {
  id: number;
  ip: string;
  /** @nullable */
  hostname?: string | null;
  status: string;
  /** @nullable */
  osGuess?: string | null;
  /** @nullable */
  macAddress?: string | null;
  openPorts: number;
  difficultyScore: HostDifficultyScore;
  hostType: HostHostType;
  services?: ServiceInfo[];
  /** @nullable */
  scanId?: number | null;
  isDemo?: boolean;
}

export type HostDetailDifficultyScore = typeof HostDetailDifficultyScore[keyof typeof HostDetailDifficultyScore];


export const HostDetailDifficultyScore = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export type HostDetailHostType = typeof HostDetailHostType[keyof typeof HostDetailHostType];


export const HostDetailHostType = {
  'domain-controller': 'domain-controller',
  server: 'server',
  workstation: 'workstation',
  unknown: 'unknown',
} as const;

export type IdentityUserPrivilegeLevel = typeof IdentityUserPrivilegeLevel[keyof typeof IdentityUserPrivilegeLevel];


export const IdentityUserPrivilegeLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  admin: 'admin',
} as const;

export type IdentityUserAccountType = typeof IdentityUserAccountType[keyof typeof IdentityUserAccountType];


export const IdentityUserAccountType = {
  standard: 'standard',
  helpdesk: 'helpdesk',
  service: 'service',
  dormant: 'dormant',
  admin: 'admin',
} as const;

export interface IdentityUser {
  id: number;
  username: string;
  privilegeLevel: IdentityUserPrivilegeLevel;
  mfaEnabled: boolean;
  accountType: IdentityUserAccountType;
  groups?: string[];
  linkedHosts?: string[];
  riskFlags?: string[];
}

export interface HostDetail {
  id: number;
  ip: string;
  /** @nullable */
  hostname?: string | null;
  status: string;
  /** @nullable */
  osGuess?: string | null;
  /** @nullable */
  macAddress?: string | null;
  openPorts: number;
  difficultyScore: HostDetailDifficultyScore;
  hostType: HostDetailHostType;
  services: ServiceInfo[];
  weaknesses: Weakness[];
  scoreReason: string;
  recommendations: string[];
  linkedUsers?: IdentityUser[];
  isDemo?: boolean;
}

export interface IdentityGroup {
  id: number;
  name: string;
  memberCount: number;
  accessLevel: string;
  members?: string[];
  permissions?: string[];
}

export type GraphNodeType = typeof GraphNodeType[keyof typeof GraphNodeType];


export const GraphNodeType = {
  host: 'host',
  service: 'service',
  user: 'user',
  group: 'group',
  admin: 'admin',
  weakness: 'weakness',
  target: 'target',
} as const;

export type GraphNodeData = { [key: string]: unknown };

export interface GraphNode {
  id: string;
  type: GraphNodeType;
  label: string;
  data?: GraphNodeData;
  /** @nullable */
  riskLevel?: string | null;
}

export interface GraphEdge {
  id: string;
  source: string;
  target: string;
  label: string;
}

export interface GraphData {
  nodes: GraphNode[];
  edges: GraphEdge[];
}

export interface AttackPathStep {
  stepNumber: number;
  actor: string;
  action: string;
  target: string;
  explanation: string;
}

export interface AttackPath {
  id: number;
  startIdentity: string;
  targetHost: string;
  steps: AttackPathStep[];
  riskSummary: string;
  controlWeaknesses: string[];
  defensiveActions: string[];
  createdAt: string;
}

export interface AttackPathSimulationInput {
  startIdentity: string;
  targetHost: string;
}

export interface ReportHost {
  ip: string;
  /** @nullable */
  hostname: string | null;
  openPorts: number;
  services: ServiceInfo[];
  weaknesses: Weakness[];
  difficultyScore: string;
}

export interface Report {
  generatedAt: string;
  scanSummary: DashboardStats;
  hosts: ReportHost[];
  attackPaths: AttackPath[];
  recommendations: string[];
}

export type ListHostsParams = {
ipFilter?: string;
hostnameFilter?: string;
serviceFilter?: string;
portFilter?: string;
scoreFilter?: ListHostsScoreFilter;
riskFilter?: ListHostsRiskFilter;
hostTypeFilter?: ListHostsHostTypeFilter;
/**
 * @nullable
 */
scanId?: number | null;
};

export type ListHostsScoreFilter = typeof ListHostsScoreFilter[keyof typeof ListHostsScoreFilter];


export const ListHostsScoreFilter = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export type ListHostsRiskFilter = typeof ListHostsRiskFilter[keyof typeof ListHostsRiskFilter];


export const ListHostsRiskFilter = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type ListHostsHostTypeFilter = typeof ListHostsHostTypeFilter[keyof typeof ListHostsHostTypeFilter];


export const ListHostsHostTypeFilter = {
  'domain-controller': 'domain-controller',
  server: 'server',
  workstation: 'workstation',
  unknown: 'unknown',
} as const;

export type GetGraphParams = {
/**
 * @nullable
 */
scanId?: number | null;
};