# lib/api-zod/src/generated/types/attackPath.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RedPath Sentinel API
 * OpenAPI spec version: 0.1.0
 */
import type { AttackPathStep } from './attackPathStep';

export interface AttackPath {
  id: number;
  startIdentity: string;
  targetHost: string;
  steps: AttackPathStep[];
  riskSummary: string;
  controlWeaknesses: string[];
  defensiveActions: string[];
  createdAt: string;
}