# lib/api-zod/src/generated/types/identityGroup.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RedPath Sentinel API
 * OpenAPI spec version: 0.1.0
 */

export interface IdentityGroup {
  id: number;
  name: string;
  memberCount: number;
  accessLevel: string;
  members?: string[];
  permissions?: string[];
}