# lib/api-zod/src/generated/types/identityUser.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * RedPath Sentinel API
 * OpenAPI spec version: 0.1.0
 */
import type { IdentityUserAccountType } from './identityUserAccountType';
import type { IdentityUserPrivilegeLevel } from './identityUserPrivilegeLevel';

export interface IdentityUser {
  id: number;
  username: string;
  privilegeLevel: IdentityUserPrivilegeLevel;
  mfaEnabled: boolean;
  accountType: IdentityUserAccountType;
  groups?: string[];
  linkedHosts?: string[];
  riskFlags?: string[];
}